refactor: split extension contract modules
