+++ src/opencode_a2a/contracts/extensions/__init__.py
@@ -1,0 +1,55 @@
+from __future__ import annotations
+
+from .capabilities import JsonRpcCapabilitySnapshot, build_capability_snapshot
+from .catalog import (
+    COMMAND_REQUEST_ALLOWED_FIELDS,
+    INTERRUPT_CALLBACK_METHODS,
+    INTERRUPT_ERROR_BUSINESS_CODES,
+    INTERRUPT_RECOVERY_METHODS,
+    PROMPT_ASYNC_REQUEST_ALLOWED_FIELDS,
+    PROVIDER_DISCOVERY_ERROR_BUSINESS_CODES,
+    PROVIDER_DISCOVERY_METHODS,
+    SESSION_METHODS,
+    SESSION_QUERY_DEFAULT_LIMIT,
+    SESSION_QUERY_ERROR_BUSINESS_CODES,
+    SESSION_QUERY_MAX_LIMIT,
+    SESSION_QUERY_PAGINATION_UNSUPPORTED,
+    SHELL_REQUEST_ALLOWED_FIELDS,
+    WORKSPACE_CONTROL_ERROR_BUSINESS_CODES,
+    WORKSPACE_CONTROL_METHODS,
+)
+from .compatibility import (
+    build_compatibility_profile_params,
+    build_protocol_compatibility_params,
+    build_service_behavior_contract_params,
+    build_wire_contract_params,
+)
+from .identifiers import (
+    ALL_EXTENSION_URIS,
+    AUTHENTICATED_ONLY_EXTENSION_URIS,
+    COMPATIBILITY_PROFILE_EXTENSION_URI,
+    EXTENSION_SPEC_DOCUMENT_PATHS_BY_URI,
+    EXTENSION_URI_NAMESPACE,
+    INTERRUPT_CALLBACK_EXTENSION_URI,
+    INTERRUPT_RECOVERY_EXTENSION_URI,
+    MODEL_SELECTION_EXTENSION_URI,
+    PROVIDER_DISCOVERY_EXTENSION_URI,
+    PUBLIC_EXTENSION_URIS,
+    SESSION_BINDING_EXTENSION_URI,
+    SESSION_MANAGEMENT_EXTENSION_URI,
+    STREAMING_EXTENSION_URI,
+    WIRE_CONTRACT_EXTENSION_URI,
+    WORKSPACE_CONTROL_EXTENSION_URI,
+)
+from .private_params import (
+    build_interrupt_recovery_extension_params,
+    build_provider_discovery_extension_params,
+    build_session_management_extension_params,
+    build_workspace_control_extension_params,
+)
+from .public_params import (
+    build_interrupt_callback_extension_params,
+    build_model_selection_extension_params,
+    build_session_binding_extension_params,
+    build_streaming_extension_params,
+)
