+++ tests/support/helpers.py
@@ -213,7 +213,13 @@ def make_request_context_with_parts(
 
 
 class DummyChatOpencodeUpstreamClient:
-    def __init__(self, settings: Settings | None = None) -> None:
+    def __init__(
+        self,
+        settings: Settings | None = None,
+        *,
+        interrupt_request_repository=None,  # noqa: ANN001
+    ) -> None:
+        del interrupt_request_repository
         self.created_sessions = 0
         self.sent_session_ids: list[str] = []
         self.sent_model_overrides: list[dict[str, str] | None] = []
